requirements_predict.txt
torch
torchvision
fastapi
pydantic
uvicorn
python-multipart
google-cloud-storage